Fix MBFD (it's working now)
But the algorithm didn't meet the expectations, probably because it
tends to distribute the VMs to various physical machines instead of
trying to put various VMs in the same PM. This strategy would work if no
VM is expected to sleep when it's in idle, but the other algorithms
(with the except of EnergyUnaware) put PMs to sleep when it's possible.

So this algorithm does not perform well, losing even to FFD, thanks to
this property. We probably need to tweak it a little, using the global
energy consumption as a base to the algorithm instead of only local EC.